refactor(handleChatModelStart): move to callbacks dir, use factory function
