+++ api/app/clients/callbacks/index.js
@@ -1,0 +1,5 @@
+const createStartHandler = require('./createStartHandler');
+
+module.exports = {
+  createStartHandler,
+};
